requirements.txt
# Python 3.12

Flask~=3.1.0
Werkzeug~=3.1.3
reportlab~=4.2.5
python-docx~=1.2.0
